rename builder into fuzzy-logic and move files
